0.9.73

+++ C/core/onefit-3.1/funcoes.h
@@ -5,7 +5,7 @@ double _exp(double x,double *par,double t,int mode,char **par_names,int nnames);
 double _x0(double x,double *par,double t,int mode);
 double _x1(double x,double *par,double t,int mode);
 double _x2(double x,double *par,double t,int mode);
-double _x3(double x,double *par,double t,double mode);
+double _x3(double x,double *par,double t,int mode);
 double _x4(double x,double *par,double t,int mode);
 double _x5(double x,double *par,double t,int mode);
 double _x6(double x,double *par,double t,int mode);
